cleaned up testing and added more features, such as syntactic sugar for operators and for loop.

+++ src/ast/FloatLitNode.java
@@ -12,7 +12,7 @@ public class FloatLitNode extends ExpressionNode {
 	
 	public FloatLitNode(String val) {
 		setVal(val);
-		setType(Scope.Type.INT);
+		setType(Scope.Type.FLOAT);
 	}
 
 	@Override
